backend/requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.0.1
bidict==0.23.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
Faker==37.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
gast==0.6.0
geoip2==5.1.0
google-pasta==0.2.0
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
libclang==18.1.1
mac-vendor-lookup==0.1.12
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
maxminddb==2.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.4.3
namex==0.0.9
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
passlib==1.7.4
priority==2.0.0
prometheus_client==0.21.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.7.1
python-http-client==3.3.7
python-jose==3.4.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.9.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9.1
scapy==2.6.1
scapy-http==1.8.2
scikit-learn==1.6.1
scipy==1.15.2
sendgrid==6.12.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tldextract==5.3.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
watchdog==6.0.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
